+++ posthog/models/cohort/cohort.py
@@ -22,21 +22,15 @@ from posthog.clickhouse.query_tagging import Feature, tag_queries
 from posthog.constants import PropertyOperatorType
 from posthog.exceptions_capture import capture_exception
 from posthog.helpers.batch_iterators import ArrayBatchIterator, BatchIterator, FunctionBatchIterator
+from posthog.models.cohort.util import delete_cohort_member, insert_cohort_members, is_person_in_cohort
 from posthog.models.file_system.file_system_mixin import FileSystemSyncMixin
 from posthog.models.file_system.file_system_representation import FileSystemRepresentation
 from posthog.models.filters.filter import Filter
 from posthog.models.person import Person, PersonDistinctId
 from posthog.models.person.person import READ_DB_FOR_PERSONS
-from posthog.models.person.util import (
-    delete_cohort_member,
-    get_person_by_uuid,
-    get_persons_by_distinct_ids,
-    insert_cohort_members,
-    is_person_in_cohort,
-)
+from posthog.models.person.util import get_person_by_uuid, get_persons_by_distinct_ids
 from posthog.models.property import Property, PropertyGroup
 from posthog.models.utils import RootTeamManager, RootTeamMixin, sane_repr
-from posthog.person_db_router import PERSONS_DB_FOR_WRITE
 from posthog.settings.base_variables import TEST
 
 if TYPE_CHECKING:
@@ -707,23 +701,21 @@ class Cohort(FileSystemSyncMixin, RootTeamMixin, models.Model):
         Returns:
             Number of batches processed.
         """
-        from posthog.models.cohort.util import insert_static_cohort
-        from posthog.models.person.util import count_cohort_members
+        from posthog.models.cohort.util import count_cohort_members, insert_static_cohort
         from posthog.personhog_client.gate import use_personhog
 
         current_batch_index = -1
         processing_error = None
-        db_write: str | None = None
+        personhog = use_personhog()
         try:
             from django.db import connections, router
 
-            db_write = router.db_for_write(Person) or "default"
-
-            if use_personhog():
-                current_batch_index = self._insert_batches_via_personhog(
-                    batch_iterator, insert_in_clickhouse, team_id=team_id
-                )
+            if personhog:
+                for batch_index, batch in batch_iterator:
+                    current_batch_index = batch_index
+                    self._insert_batch_via_personhog(batch, insert_in_clickhouse, team_id=team_id)
             else:
+                db_write = router.db_for_write(Person) or "default"
                 db_read = router.db_for_read(Person) or "default"
                 persons_connection = connections[db_write]
                 cursor = persons_connection.cursor()
@@ -735,6 +727,11 @@ class Cohort(FileSystemSyncMixin, RootTeamMixin, models.Model):
                         Person.objects.db_manager(db_read).filter(team_id=team_id).filter(uuid__in=batch)
                     )
                     if insert_in_clickhouse:
+                        # Both querysets must use db_write so Django can merge the
+                        # .exclude() into a single NOT IN subquery. Using db_read
+                        # for Person + db_write for CohortPeople causes a
+                        # "Subqueries aren't allowed across different databases"
+                        # ValueError when the aliases differ (production config).
                         insert_uuids_query = (
                             Person.objects.using(db_write)  # nosemgrep: no-direct-persons-db-orm
                             .filter(team_id=team_id, uuid__in=batch)
@@ -750,6 +747,10 @@ class Cohort(FileSystemSyncMixin, RootTeamMixin, models.Model):
                             team_id=team_id,
                         )
 
+                    # Dedup via LEFT JOIN so the exclusion stays entirely in SQL,
+                    # avoiding the O(cohort_size) memory cost of loading all
+                    # existing member IDs into Python. Both tables live on the
+                    # persons DB so the join works on the db_write cursor.
                     sql, params = persons_query.distinct("pk").only("pk").query.sql_with_params()
                     query = f"""
                         INSERT INTO "{cohort_people_table}" ("person_id", "cohort_id", "version")
@@ -775,10 +776,13 @@ class Cohort(FileSystemSyncMixin, RootTeamMixin, models.Model):
                 },
             )
         finally:
+            # Always update the count and cohort state, even if processing failed
             try:
-                count = count_cohort_members(cohort_id=self.id, team_id=self.team_id, using_database=db_write)
+                count = count_cohort_members(cohort_id=self.id, team_id=self.team_id, consistency="strong")
                 self.count = count
             except Exception as count_err:
+                # If count calculation fails, log the error but don't override the processing error.
+                # Leave existing count unchanged - it's better than None.
                 logger.exception(
                     "Failed to calculate static cohort size",
                     cohort_id=self.id,
@@ -793,47 +797,40 @@ class Cohort(FileSystemSyncMixin, RootTeamMixin, models.Model):
 
         return current_batch_index + 1
 
-    def _insert_batches_via_personhog(
+    def _insert_batch_via_personhog(
         self,
-        batch_iterator: BatchIterator[str],
+        batch: list[str],
         insert_in_clickhouse: bool,
         *,
         team_id: int,
-    ) -> int:
-        """Personhog path for batch cohort member insertion.
+    ) -> None:
+        """Personhog path for inserting a single batch of cohort members.
 
         Resolves UUIDs → person IDs via personhog, then calls the
         InsertCohortMembers RPC. ClickHouse inserts (if requested)
         exclude persons already in the cohort because the
         person_static_cohort table's ORDER BY includes a per-row UUID,
         preventing ReplacingMergeTree from deduplicating repeated inserts.
-
-        Returns the last batch index processed.
         """
         from posthog.models.cohort.util import insert_static_cohort
         from posthog.models.person.sql import PERSON_STATIC_COHORT_TABLE
         from posthog.models.person.util import get_persons_by_uuids
 
-        last_batch_index = -1
-        for batch_index, batch in batch_iterator:
-            last_batch_index = batch_index
-            persons = get_persons_by_uuids(self.team, batch)
-            if not persons:
-                continue
+        persons = get_persons_by_uuids(self.team, batch)
+        if not persons:
+            return
 
-            person_ids = [p.id for p in persons]
-            person_uuids = [p.uuid for p in persons]
+        person_ids = [p.id for p in persons]
+        person_uuids = [p.uuid for p in persons]
 
-            if insert_in_clickhouse:
-                uuid_strs = [str(u) for u in person_uuids]
-                existing_uuids = self._get_existing_ch_member_uuids(uuid_strs, team_id, PERSON_STATIC_COHORT_TABLE)
-                new_uuids = [u for u in person_uuids if str(u) not in existing_uuids]
-                if new_uuids:
-                    insert_static_cohort(new_uuids, self.pk, team_id=team_id)
+        if insert_in_clickhouse:
+            uuid_strs = [str(u) for u in person_uuids]
+            existing_uuids = self._get_existing_ch_member_uuids(uuid_strs, team_id, PERSON_STATIC_COHORT_TABLE)
+            new_uuids = [u for u in person_uuids if str(u) not in existing_uuids]
+            if new_uuids:
+                insert_static_cohort(new_uuids, self.pk, team_id=team_id)
 
-            insert_cohort_members(team_id, self.pk, person_ids, self.version)
-
-        return last_batch_index
+        insert_cohort_members(team_id, self.pk, person_ids, self.version)
 
     def _get_existing_ch_member_uuids(
         self,
@@ -899,12 +896,11 @@ class Cohort(FileSystemSyncMixin, RootTeamMixin, models.Model):
             # data exists in CH but not PG due to past sync issues
             remove_person_from_static_cohort(person.uuid, self.pk, team_id=team_id)
 
-            # Update count - use write database to avoid replication lag after delete
             try:
                 count = get_static_cohort_size(
                     cohort_id=self.id,
                     team_id=team_id,
-                    using_database=PERSONS_DB_FOR_WRITE,
+                    consistency="strong",
                 )
                 self.count = count
                 self.save(update_fields=["count"])
@@ -1113,11 +1109,10 @@ def cohort_people_changed(sender, instance: "CohortPeople", **kwargs):
         person_uuid = instance.person_id
 
         cohort = Cohort.objects.get(id=cohort_id)
-        # Use write database to avoid replication lag after delete
         cohort.count = get_static_cohort_size(
             cohort_id=cohort.id,
             team_id=cohort.team_id,
-            using_database=PERSONS_DB_FOR_WRITE,
+            consistency="strong",
         )
 
         # Clear cohort_type if count exceeds the realtime threshold
