+++ posthog/models/cohort/test/test_cohort_personhog.py
@@ -108,7 +108,9 @@ class TestRemoveUserByUuid(PersonhogTestMixin, BaseTest):
         result = cohort.remove_user_by_uuid(str(person.uuid), team_id=self.team.id)
 
         assert result is True
-        if not self.personhog:
+        if self.personhog:
+            assert (cohort.id, person.id) not in self._fake_client._cohort_members
+        else:
             assert not CohortPeople.objects.filter(cohort=cohort, person=person).exists()
         mock_remove_ch.assert_called_once()
         call_args = mock_remove_ch.call_args
@@ -191,7 +193,9 @@ class TestRemoveUserByUuid(PersonhogTestMixin, BaseTest):
         result = cohort.remove_user_by_uuid(str(person.uuid), team_id=self.team.id)
 
         assert result is True
-        if not self.personhog:
+        if self.personhog:
+            assert (cohort.id, person.id) not in self._fake_client._cohort_members
+        else:
             assert not CohortPeople.objects.filter(cohort=cohort, person=person).exists()
         mock_remove_ch.assert_called_once()
 
@@ -209,7 +213,7 @@ class TestRemoveUserByUuid(PersonhogTestMixin, BaseTest):
 @parameterized_class(("personhog",), [(False,), (True,)])
 class TestCheckCohortMembership(PersonhogTestMixin, BaseTest):
     def test_returns_true_for_member(self):
-        from posthog.models.person.util import check_cohort_membership, is_person_in_cohort
+        from posthog.models.cohort.util import check_cohort_membership, is_person_in_cohort
 
         person = self._seed_person(team=self.team, distinct_ids=["d1"])
         cohort = Cohort.objects.create(team=self.team, groups=[], is_static=True, name="c1")
@@ -221,7 +225,7 @@ class TestCheckCohortMembership(PersonhogTestMixin, BaseTest):
         self._assert_personhog_called("check_cohort_membership")
 
     def test_returns_false_for_non_member(self):
-        from posthog.models.person.util import is_person_in_cohort
+        from posthog.models.cohort.util import is_person_in_cohort
 
         person = self._seed_person(team=self.team, distinct_ids=["d1"])
         cohort = Cohort.objects.create(team=self.team, groups=[], is_static=True, name="c1")
@@ -229,7 +233,7 @@ class TestCheckCohortMembership(PersonhogTestMixin, BaseTest):
         assert is_person_in_cohort(team_id=self.team.id, person_id=person.id, cohort_id=cohort.id) is False
 
     def test_returns_empty_dict_for_empty_cohort_ids(self):
-        from posthog.models.person.util import check_cohort_membership
+        from posthog.models.cohort.util import check_cohort_membership
 
         person = self._seed_person(team=self.team, distinct_ids=["d1"])
 
@@ -237,7 +241,7 @@ class TestCheckCohortMembership(PersonhogTestMixin, BaseTest):
         self._assert_personhog_not_called("check_cohort_membership")
 
     def test_mixed_membership(self):
-        from posthog.models.person.util import check_cohort_membership
+        from posthog.models.cohort.util import check_cohort_membership
 
         person = self._seed_person(team=self.team, distinct_ids=["d1"])
         c1 = Cohort.objects.create(team=self.team, groups=[], is_static=True, name="c1")
@@ -257,7 +261,7 @@ class TestCheckCohortMembership(PersonhogTestMixin, BaseTest):
         of which path (personhog vs ORM) runs. Tenant scoping happens in the
         public wrapper before dispatch, so the RPC is never called for
         out-of-team cohort_ids."""
-        from posthog.models.person.util import check_cohort_membership
+        from posthog.models.cohort.util import check_cohort_membership
 
         other_team = Team.objects.create(organization=self.organization)
         person = self._seed_person(team=self.team, distinct_ids=["d1"])
@@ -277,7 +281,7 @@ class TestCheckCohortMembership(PersonhogTestMixin, BaseTest):
         """When a mix of in-team and out-of-team cohort_ids is passed, only the
         in-team ones reach downstream; the caller still gets a dict keyed by
         every requested id with out-of-team ones reported as non-member."""
-        from posthog.models.person.util import check_cohort_membership
+        from posthog.models.cohort.util import check_cohort_membership
 
         other_team = Team.objects.create(organization=self.organization)
         person = self._seed_person(team=self.team, distinct_ids=["d1"])
@@ -297,7 +301,7 @@ class TestCheckCohortMembershipFallback(BaseTest):
     """Routing test: verifies ORM fallback when the personhog gate is disabled."""
 
     def test_falls_back_to_orm_when_personhog_disabled(self):
-        from posthog.models.person.util import is_person_in_cohort
+        from posthog.models.cohort.util import is_person_in_cohort
 
         person = Person.objects.create(team=self.team, distinct_ids=["d1"])
         cohort = Cohort.objects.create(team=self.team, groups=[], is_static=True, name="c1")
@@ -395,7 +399,7 @@ class TestPropertyToQStaticCohortShortCircuit(PersonhogTestMixin, BaseTest):
 @parameterized_class(("personhog",), [(False,), (True,)])
 class TestListCohortMemberIds(PersonhogTestMixin, BaseTest):
     def test_returns_member_ids(self):
-        from posthog.models.person.util import list_cohort_member_ids
+        from posthog.models.cohort.util import list_cohort_member_ids
 
         p1 = self._seed_person(team=self.team, distinct_ids=["d1"])
         p2 = self._seed_person(team=self.team, distinct_ids=["d2"])
@@ -411,7 +415,7 @@ class TestListCohortMemberIds(PersonhogTestMixin, BaseTest):
         self._assert_personhog_called("list_cohort_member_ids")
 
     def test_returns_empty_for_empty_cohort(self):
-        from posthog.models.person.util import list_cohort_member_ids
+        from posthog.models.cohort.util import list_cohort_member_ids
 
         cohort = Cohort.objects.create(team=self.team, groups=[], is_static=True, name="c1")
 
@@ -420,7 +424,7 @@ class TestListCohortMemberIds(PersonhogTestMixin, BaseTest):
         assert result == []
 
     def test_excludes_non_members(self):
-        from posthog.models.person.util import list_cohort_member_ids
+        from posthog.models.cohort.util import list_cohort_member_ids
 
         p1 = self._seed_person(team=self.team, distinct_ids=["d1"])
         p2 = self._seed_person(team=self.team, distinct_ids=["d2"])
@@ -436,7 +440,7 @@ class TestListCohortMemberIds(PersonhogTestMixin, BaseTest):
         assert result == [p1.id]
 
     def test_isolates_cross_team_cohort(self):
-        from posthog.models.person.util import list_cohort_member_ids
+        from posthog.models.cohort.util import list_cohort_member_ids
 
         other_team = Team.objects.create(organization=self.organization)
         person = self._seed_person(team=other_team, distinct_ids=["d1"])
@@ -452,7 +456,7 @@ class TestListCohortMemberIds(PersonhogTestMixin, BaseTest):
 
 class TestListCohortMemberIdsFallback(BaseTest):
     def test_falls_back_to_orm_when_personhog_disabled(self):
-        from posthog.models.person.util import list_cohort_member_ids
+        from posthog.models.cohort.util import list_cohort_member_ids
 
         person = Person.objects.create(team=self.team, distinct_ids=["d1"])
         cohort = Cohort.objects.create(team=self.team, groups=[], is_static=True, name="c1")
@@ -538,7 +542,7 @@ class TestInsertCohortMembers(PersonhogTestMixin, BaseTest):
         return Cohort.objects.create(team=self.team, groups=[], is_static=True, name="test cohort")
 
     def test_inserts_members(self):
-        from posthog.models.person.util import insert_cohort_members
+        from posthog.models.cohort.util import insert_cohort_members
 
         p1 = self._seed_person(team=self.team, distinct_ids=["d1"])
         p2 = self._seed_person(team=self.team, distinct_ids=["d2"])
@@ -547,12 +551,15 @@ class TestInsertCohortMembers(PersonhogTestMixin, BaseTest):
         inserted = insert_cohort_members(self.team.id, cohort.id, [p1.id, p2.id], version=1)
 
         assert inserted > 0
-        if not self.personhog:
+        if self.personhog:
+            assert (cohort.id, p1.id) in self._fake_client._cohort_members
+            assert (cohort.id, p2.id) in self._fake_client._cohort_members
+        else:
             assert CohortPeople.objects.filter(cohort=cohort).count() == 2
         self._assert_personhog_called("insert_cohort_members")
 
     def test_deduplicates_existing_members(self):
-        from posthog.models.person.util import insert_cohort_members
+        from posthog.models.cohort.util import insert_cohort_members
 
         person = self._seed_person(team=self.team, distinct_ids=["d1"])
         cohort = self._create_static_cohort()
@@ -567,7 +574,7 @@ class TestInsertCohortMembers(PersonhogTestMixin, BaseTest):
             assert CohortPeople.objects.filter(cohort=cohort).count() == 1
 
     def test_returns_zero_for_empty_list(self):
-        from posthog.models.person.util import insert_cohort_members
+        from posthog.models.cohort.util import insert_cohort_members
 
         cohort = self._create_static_cohort()
 
@@ -575,7 +582,7 @@ class TestInsertCohortMembers(PersonhogTestMixin, BaseTest):
         self._assert_personhog_not_called("insert_cohort_members")
 
     def test_rejects_cross_team_cohort(self):
-        from posthog.models.person.util import insert_cohort_members
+        from posthog.models.cohort.util import insert_cohort_members
 
         other_team = Team.objects.create(organization=self.organization)
         person = self._seed_person(team=self.team, distinct_ids=["d1"])
@@ -590,7 +597,7 @@ class TestInsertCohortMembers(PersonhogTestMixin, BaseTest):
 
 class TestInsertCohortMembersFallback(BaseTest):
     def test_falls_back_to_orm_when_personhog_disabled(self):
-        from posthog.models.person.util import insert_cohort_members
+        from posthog.models.cohort.util import insert_cohort_members
 
         person = Person.objects.create(team=self.team, distinct_ids=["d1"])
         cohort = Cohort.objects.create(team=self.team, groups=[], is_static=True, name="c1")
@@ -610,7 +617,7 @@ class TestDeleteCohortMember(PersonhogTestMixin, BaseTest):
         return Cohort.objects.create(team=self.team, groups=[], is_static=True, name="test cohort")
 
     def test_deletes_existing_member(self):
-        from posthog.models.person.util import delete_cohort_member
+        from posthog.models.cohort.util import delete_cohort_member
 
         person = self._seed_person(team=self.team, distinct_ids=["d1"])
         cohort = self._create_static_cohort()
@@ -625,7 +632,7 @@ class TestDeleteCohortMember(PersonhogTestMixin, BaseTest):
         self._assert_personhog_called("delete_cohort_member")
 
     def test_returns_false_for_non_member(self):
-        from posthog.models.person.util import delete_cohort_member
+        from posthog.models.cohort.util import delete_cohort_member
 
         person = self._seed_person(team=self.team, distinct_ids=["d1"])
         cohort = self._create_static_cohort()
@@ -635,7 +642,7 @@ class TestDeleteCohortMember(PersonhogTestMixin, BaseTest):
         assert result is False
 
     def test_rejects_cross_team_cohort(self):
-        from posthog.models.person.util import delete_cohort_member
+        from posthog.models.cohort.util import delete_cohort_member
 
         other_team = Team.objects.create(organization=self.organization)
         person = self._seed_person(team=other_team, distinct_ids=["d1"])
@@ -652,7 +659,7 @@ class TestDeleteCohortMember(PersonhogTestMixin, BaseTest):
 
 class TestDeleteCohortMemberFallback(BaseTest):
     def test_falls_back_to_orm_when_personhog_disabled(self):
-        from posthog.models.person.util import delete_cohort_member
+        from posthog.models.cohort.util import delete_cohort_member
 
         person = Person.objects.create(team=self.team, distinct_ids=["d1"])
         cohort = Cohort.objects.create(team=self.team, groups=[], is_static=True, name="c1")
@@ -671,7 +678,7 @@ class TestDeleteCohortMembersBulk(PersonhogTestMixin, BaseTest):
     CLASS_DATA_LEVEL_SETUP = False
 
     def test_deletes_all_members_for_cohorts(self):
-        from posthog.models.person.util import delete_cohort_members_bulk
+        from posthog.models.cohort.util import delete_cohort_members_bulk
 
         p1 = self._seed_person(team=self.team, distinct_ids=["d1"])
         p2 = self._seed_person(team=self.team, distinct_ids=["d2"])
@@ -690,7 +697,7 @@ class TestDeleteCohortMembersBulk(PersonhogTestMixin, BaseTest):
         self._assert_personhog_called("delete_cohort_members_bulk")
 
     def test_returns_zero_for_empty_list(self):
-        from posthog.models.person.util import delete_cohort_members_bulk
+        from posthog.models.cohort.util import delete_cohort_members_bulk
 
         assert delete_cohort_members_bulk(self.team.id, []) == 0
         self._assert_personhog_not_called("delete_cohort_members_bulk")
@@ -698,7 +705,7 @@ class TestDeleteCohortMembersBulk(PersonhogTestMixin, BaseTest):
 
 class TestDeleteCohortMembersBulkFallback(BaseTest):
     def test_falls_back_to_orm_when_personhog_disabled(self):
-        from posthog.models.person.util import delete_cohort_members_bulk
+        from posthog.models.cohort.util import delete_cohort_members_bulk
 
         person = Person.objects.create(team=self.team, distinct_ids=["d1"])
         cohort = Cohort.objects.create(team=self.team, groups=[], is_static=True, name="c1")
@@ -716,7 +723,7 @@ class TestDeleteCohortMembersBulkMaxIterations(BaseTest):
     def test_stops_after_max_iterations(self):
         from unittest.mock import MagicMock
 
-        from posthog.models.person import util as person_util
+        from posthog.models.cohort import util as cohort_util
         from posthog.personhog_client.proto import DeleteCohortMembersBulkResponse
 
         mock_client = MagicMock()
@@ -724,10 +731,10 @@ class TestDeleteCohortMembersBulkMaxIterations(BaseTest):
 
         max_iters = 5
         with (
-            patch.object(person_util, "_DELETE_BULK_MAX_ITERATIONS", max_iters),
+            patch.object(cohort_util, "_DELETE_BULK_MAX_ITERATIONS", max_iters),
             patch("posthog.personhog_client.client.get_personhog_client", return_value=mock_client),
         ):
-            total = person_util._delete_cohort_members_bulk_via_personhog([1], batch_size=100)
+            total = cohort_util._delete_cohort_members_bulk_via_personhog([1], batch_size=100)
 
         assert mock_client.delete_cohort_members_bulk.call_count == max_iters
         assert total == 100 * max_iters
@@ -736,7 +743,7 @@ class TestDeleteCohortMembersBulkMaxIterations(BaseTest):
 @parameterized_class(("personhog",), [(False,), (True,)])
 class TestCountCohortMembers(PersonhogTestMixin, BaseTest):
     def test_returns_count(self):
-        from posthog.models.person.util import count_cohort_members
+        from posthog.models.cohort.util import count_cohort_members
 
         p1 = self._seed_person(team=self.team, distinct_ids=["d1"])
         p2 = self._seed_person(team=self.team, distinct_ids=["d2"])
@@ -750,14 +757,14 @@ class TestCountCohortMembers(PersonhogTestMixin, BaseTest):
         self._assert_personhog_called("count_cohort_members")
 
     def test_returns_zero_for_empty_cohort(self):
-        from posthog.models.person.util import count_cohort_members
+        from posthog.models.cohort.util import count_cohort_members
 
         cohort = Cohort.objects.create(team=self.team, groups=[], is_static=True, name="c1")
 
         assert count_cohort_members(self.team.id, cohort.id) == 0
 
     def test_isolates_cross_team_cohort(self):
-        from posthog.models.person.util import count_cohort_members
+        from posthog.models.cohort.util import count_cohort_members
 
         other_team = Team.objects.create(organization=self.organization)
         person = self._seed_person(team=other_team, distinct_ids=["d1"])
@@ -771,7 +778,7 @@ class TestCountCohortMembers(PersonhogTestMixin, BaseTest):
 
 class TestCountCohortMembersFallback(BaseTest):
     def test_falls_back_to_orm_when_personhog_disabled(self):
-        from posthog.models.person.util import count_cohort_members
+        from posthog.models.cohort.util import count_cohort_members
 
         person = Person.objects.create(team=self.team, distinct_ids=["d1"])
         cohort = Cohort.objects.create(team=self.team, groups=[], is_static=True, name="c1")
