+++ posthog/models/cohort/util.py
@@ -53,6 +53,15 @@ from posthog.models.person.sql import (
     PERSON_STATIC_COHORT_TABLE,
 )
 from posthog.models.property import Property, PropertyGroup
+from posthog.personhog_client import ReadConsistency, consistency_to_read_options
+from posthog.personhog_client.proto import (
+    CheckCohortMembershipRequest,
+    CountCohortMembersRequest,
+    DeleteCohortMemberRequest,
+    DeleteCohortMembersBulkRequest,
+    InsertCohortMembersRequest,
+    ListCohortMemberIdsRequest,
+)
 from posthog.queries.person_distinct_id_query import get_team_distinct_ids_query
 from posthog.queries.util import PersonPropertiesMode
 
@@ -635,10 +644,13 @@ def remove_person_from_static_cohort(person_uuid: uuid.UUID, cohort_id: int, *, 
     )
 
 
-def get_static_cohort_size(*, cohort_id: int, team_id: int, using_database: str | None = None) -> int:
-    from posthog.models.person.util import count_cohort_members
-
-    return count_cohort_members(team_id=team_id, cohort_id=cohort_id, using_database=using_database)
+def get_static_cohort_size(
+    *,
+    cohort_id: int,
+    team_id: int,
+    consistency: ReadConsistency = "eventual",
+) -> int:
+    return count_cohort_members(team_id=team_id, cohort_id=cohort_id, consistency=consistency)
 
 
 def recalculate_cohortpeople(
@@ -1110,3 +1122,331 @@ def insert_cohort_people_into_pg(cohort: "Cohort", *, team_id: int):
         fetch_batch, CH_PAGE_SIZE, initial_cursor="00000000-0000-0000-0000-000000000000"
     )
     cohort._insert_users_list_with_batching(batch_iterator, insert_in_clickhouse=False, team_id=team_id)
+
+
+# ── Cohort membership operations (Postgres / personhog) ───────────────
+
+
+def _check_cohort_membership_via_personhog(person_id: int, cohort_ids: list[int]) -> dict[int, bool]:
+    from posthog.personhog_client.client import get_personhog_client
+
+    client = get_personhog_client()
+    if client is None:
+        raise RuntimeError("personhog client not configured")
+
+    resp = client.check_cohort_membership(CheckCohortMembershipRequest(person_id=person_id, cohort_ids=cohort_ids))
+    membership_by_cohort: dict[int, bool] = {m.cohort_id: m.is_member for m in resp.memberships}
+    return {cohort_id: membership_by_cohort.get(cohort_id, False) for cohort_id in cohort_ids}
+
+
+def check_cohort_membership(team_id: int, person_id: int, cohort_ids: list[int]) -> dict[int, bool]:
+    """Return ``{cohort_id: is_member}`` for the given person.
+
+    Routes through personhog when the gate is enabled, falling back to a Django
+    ORM query against ``posthog_cohortpeople`` (on the persons DB) otherwise.
+    Membership for cohorts the person is not in is returned as ``False`` rather
+    than being omitted, so callers can index directly by cohort_id.
+    """
+    from posthog.models.person.util import _personhog_routed
+
+    if not cohort_ids:
+        return {}
+
+    # Local import to avoid circulars (cohort → person via CohortPeople FK).
+    from posthog.models.cohort.cohort import Cohort, CohortPeople
+    from posthog.models.person.person import READ_DB_FOR_PERSONS
+
+    # Scope cohort_ids to the team via Cohort on the default DB before querying
+    # either the personhog RPC or the persons-DB CohortPeople table. Neither
+    # downstream path enforces team ownership (posthog_cohortpeople has no
+    # team_id column; the RPC just filters by person_id + cohort_id), so the
+    # tenant boundary has to be applied here. Cohorts belonging to a different
+    # team are reported as ``False`` rather than looked up. Same pattern as
+    # `posthog.models.team.util.delete_bulky_postgres_data`.
+    scoped_cohort_ids = list(Cohort.objects.filter(id__in=cohort_ids, team_id=team_id).values_list("id", flat=True))
+    if not scoped_cohort_ids:
+        return dict.fromkeys(cohort_ids, False)
+
+    def orm_fn() -> dict[int, bool]:
+        member_ids = set(
+            CohortPeople.objects.db_manager(READ_DB_FOR_PERSONS)
+            .filter(person_id=person_id, cohort_id__in=scoped_cohort_ids)
+            .values_list("cohort_id", flat=True)
+        )
+        return {cohort_id: cohort_id in member_ids for cohort_id in scoped_cohort_ids}
+
+    scoped_result = _personhog_routed(
+        "check_cohort_membership",
+        lambda: _check_cohort_membership_via_personhog(person_id, scoped_cohort_ids),
+        orm_fn,
+        team_id=team_id,
+    )
+    # Expand back to the caller's original cohort_ids; cohorts that were scoped
+    # out (not owned by this team) register as non-member.
+    return {cohort_id: scoped_result.get(cohort_id, False) for cohort_id in cohort_ids}
+
+
+def is_person_in_cohort(team_id: int, person_id: int, cohort_id: int) -> bool:
+    """Convenience single-cohort variant of ``check_cohort_membership``."""
+    return check_cohort_membership(team_id, person_id, [cohort_id]).get(cohort_id, False)
+
+
+_LIST_COHORT_MEMBER_IDS_PAGE_SIZE = 10_000
+
+
+def _list_cohort_member_ids_via_personhog(cohort_id: int) -> list[int]:
+    from posthog.personhog_client.client import get_personhog_client
+
+    client = get_personhog_client()
+    if client is None:
+        raise RuntimeError("personhog client not configured")
+
+    all_ids: list[int] = []
+    cursor = 0
+    while True:
+        resp = client.list_cohort_member_ids(
+            ListCohortMemberIdsRequest(cohort_id=cohort_id, cursor=cursor, limit=_LIST_COHORT_MEMBER_IDS_PAGE_SIZE)
+        )
+        all_ids.extend(resp.person_ids)
+        if resp.next_cursor == 0:
+            break
+        cursor = resp.next_cursor
+    return all_ids
+
+
+def list_cohort_member_ids(team_id: int, cohort_id: int) -> list[int]:
+    """Return all person IDs belonging to a static cohort.
+
+    Routes through personhog when the gate is enabled, falling back to a Django
+    ORM query against ``posthog_cohortpeople`` (on the persons DB) otherwise.
+    """
+    from posthog.models.cohort.cohort import Cohort, CohortPeople
+    from posthog.models.person.person import READ_DB_FOR_PERSONS
+    from posthog.models.person.util import _personhog_routed
+
+    # Validate cohort ownership on the default DB before querying the persons DB
+    # or the personhog RPC — neither downstream path enforces team isolation
+    # (posthog_cohortpeople has no team_id column; the proto has no team_id field).
+    # Same pattern as check_cohort_membership / delete_bulky_postgres_data.
+    if not Cohort.objects.filter(id=cohort_id, team_id=team_id).exists():
+        return []
+
+    def orm_fn() -> list[int]:
+        return list(
+            CohortPeople.objects.db_manager(READ_DB_FOR_PERSONS)
+            .filter(cohort_id=cohort_id)
+            .values_list("person_id", flat=True)
+        )
+
+    return _personhog_routed(
+        "list_cohort_member_ids",
+        lambda: _list_cohort_member_ids_via_personhog(cohort_id),
+        orm_fn,
+        team_id=team_id,
+    )
+
+
+# ── Cohort write operations ─────────────────────────────────────────
+
+
+def _insert_cohort_members_via_personhog(cohort_id: int, person_ids: list[int], version: int | None) -> int:
+    from posthog.personhog_client.client import get_personhog_client
+
+    client = get_personhog_client()
+    if client is None:
+        raise RuntimeError("personhog client not configured")
+
+    resp = client.insert_cohort_members(
+        InsertCohortMembersRequest(
+            cohort_id=cohort_id,
+            person_ids=person_ids,
+            version=version,
+        )
+    )
+    return resp.inserted_count
+
+
+def insert_cohort_members(team_id: int, cohort_id: int, person_ids: list[int], version: int | None) -> int:
+    """Insert person IDs into a static cohort's PG membership table.
+
+    Routes through personhog when the gate is enabled, falling back to a raw
+    SQL INSERT against ``posthog_cohortpeople`` (on the persons DB) otherwise.
+    Returns the number of newly inserted rows (duplicates are skipped).
+    """
+    from posthog.models.person.util import _personhog_routed
+
+    if not person_ids:
+        return 0
+
+    from posthog.models.cohort.cohort import Cohort, CohortPeople
+    from posthog.models.person import Person
+
+    if not Cohort.objects.filter(id=cohort_id, team_id=team_id).exists():
+        return 0
+
+    def orm_fn() -> int:
+        from django.db import connections, router
+
+        db_write = router.db_for_write(Person) or "default"
+        cursor = connections[db_write].cursor()
+        table = CohortPeople._meta.db_table
+        placeholders = ",".join(["%s"] * len(person_ids))
+        query = f"""
+            INSERT INTO "{table}" ("person_id", "cohort_id", "version")
+            SELECT pid, %s, %s
+            FROM UNNEST(ARRAY[{placeholders}]) AS t(pid)
+            WHERE NOT EXISTS (
+                SELECT 1 FROM "{table}" AS cp
+                WHERE cp."person_id" = pid AND cp."cohort_id" = %s
+            )
+        """
+        cursor.execute(query, [cohort_id, version, *person_ids, cohort_id])
+        return cursor.rowcount
+
+    return _personhog_routed(
+        "insert_cohort_members",
+        lambda: _insert_cohort_members_via_personhog(cohort_id, person_ids, version),
+        orm_fn,
+        team_id=team_id,
+    )
+
+
+def _delete_cohort_member_via_personhog(cohort_id: int, person_id: int) -> bool:
+    from posthog.personhog_client.client import get_personhog_client
+
+    client = get_personhog_client()
+    if client is None:
+        raise RuntimeError("personhog client not configured")
+
+    resp = client.delete_cohort_member(DeleteCohortMemberRequest(cohort_id=cohort_id, person_id=person_id))
+    return resp.deleted
+
+
+def delete_cohort_member(team_id: int, cohort_id: int, person_id: int) -> bool:
+    """Remove a single person from a static cohort's PG membership table.
+
+    Routes through personhog when the gate is enabled, falling back to a Django
+    ORM DELETE against ``posthog_cohortpeople`` (on the persons DB) otherwise.
+    Returns ``True`` if a row was deleted, ``False`` otherwise.
+    """
+    from posthog.models.cohort.cohort import Cohort, CohortPeople
+    from posthog.models.person.util import _personhog_routed
+
+    if not Cohort.objects.filter(id=cohort_id, team_id=team_id).exists():
+        return False
+
+    def orm_fn() -> bool:
+        deleted_count, _ = CohortPeople.objects.filter(cohort_id=cohort_id, person_id=person_id).delete()
+        return deleted_count > 0
+
+    return _personhog_routed(
+        "delete_cohort_member",
+        lambda: _delete_cohort_member_via_personhog(cohort_id, person_id),
+        orm_fn,
+        team_id=team_id,
+    )
+
+
+_DELETE_BULK_MAX_ITERATIONS = 10_000
+
+
+def _delete_cohort_members_bulk_via_personhog(cohort_ids: list[int], batch_size: int) -> int:
+    from posthog.personhog_client.client import get_personhog_client
+
+    client = get_personhog_client()
+    if client is None:
+        raise RuntimeError("personhog client not configured")
+
+    total_deleted = 0
+    for _ in range(_DELETE_BULK_MAX_ITERATIONS):
+        resp = client.delete_cohort_members_bulk(
+            DeleteCohortMembersBulkRequest(cohort_ids=cohort_ids, batch_size=batch_size)
+        )
+        total_deleted += resp.deleted_count
+        if resp.deleted_count < batch_size:
+            break
+    else:
+        logger.error(
+            "delete_cohort_members_bulk_max_iterations",
+            cohort_ids=cohort_ids,
+            total_deleted=total_deleted,
+            max_iterations=_DELETE_BULK_MAX_ITERATIONS,
+        )
+    return total_deleted
+
+
+def delete_cohort_members_bulk(team_id: int, cohort_ids: list[int], batch_size: int = 10_000) -> int:
+    """Delete all PG cohort membership rows for the given cohort IDs.
+
+    Routes through personhog when the gate is enabled, falling back to the
+    Django ORM ``_raw_delete`` against ``posthog_cohortpeople`` otherwise.
+    Returns the total number of deleted rows.
+    """
+    from posthog.models.person.util import _personhog_routed
+
+    if not cohort_ids:
+        return 0
+
+    from posthog.models.cohort.cohort import CohortPeople
+
+    def orm_fn() -> int:
+        from django.db import router
+
+        qs = CohortPeople.objects.filter(cohort_id__in=cohort_ids)
+        db_alias = router.db_for_write(CohortPeople)
+        return qs._raw_delete(db_alias)
+
+    return _personhog_routed(
+        "delete_cohort_members_bulk",
+        lambda: _delete_cohort_members_bulk_via_personhog(cohort_ids, batch_size),
+        orm_fn,
+        team_id=team_id,
+    )
+
+
+def _count_cohort_members_via_personhog(cohort_ids: list[int], consistency: ReadConsistency) -> int:
+    from posthog.personhog_client.client import get_personhog_client
+
+    client = get_personhog_client()
+    if client is None:
+        raise RuntimeError("personhog client not configured")
+
+    resp = client.count_cohort_members(
+        CountCohortMembersRequest(
+            cohort_ids=cohort_ids,
+            read_options=consistency_to_read_options(consistency),
+        )
+    )
+    return resp.count
+
+
+def count_cohort_members(team_id: int, cohort_id: int, *, consistency: ReadConsistency = "eventual") -> int:
+    """Return the number of persons in a static cohort.
+
+    Routes through personhog when the gate is enabled, falling back to a Django
+    ORM COUNT against ``posthog_cohortpeople`` (on the persons DB) otherwise.
+
+    Use ``consistency="strong"`` when reading immediately after a write
+    (e.g. after inserting or deleting cohort members) to avoid stale counts
+    from replication lag (ORM) or eventual consistency (personhog).
+    """
+    from posthog.models.cohort.cohort import Cohort, CohortPeople
+    from posthog.models.person.util import _personhog_routed
+
+    if not Cohort.objects.filter(id=cohort_id, team_id=team_id).exists():
+        return 0
+
+    def orm_fn() -> int:
+        qs = CohortPeople.objects.filter(cohort_id=cohort_id, person__team_id=team_id)
+        if consistency == "strong":
+            from posthog.person_db_router import PERSONS_DB_FOR_WRITE
+
+            qs = qs.using(PERSONS_DB_FOR_WRITE)
+        return qs.count()
+
+    return _personhog_routed(
+        "count_cohort_members",
+        lambda: _count_cohort_members_via_personhog([cohort_id], consistency),
+        orm_fn,
+        team_id=team_id,
+    )
